change mycourse layout
