+++ src/app/(tabs)/myCourses/_layout.tsx
@@ -1,0 +1,11 @@
+import { Stack } from 'expo-router';
+import React from 'react';
+
+export default function Layout() {
+  return (
+    <Stack>
+      <Stack.Screen name="index" options={{ headerShown: false }} />
+      <Stack.Screen name="form/[link]" options={{ title: 'Form Viewer' }} />
+    </Stack>
+  );
+}
