+++ src/constants/url.ts
@@ -1,4 +1,5 @@
 const apiBaseURL = "http://seiedutrust-001-site2.gtempurl.com"
+const apiBaseURL2 = "https://www.seiedutrust.com/Home/ShowForm?id=8650"
 
 export const facultyMembersURL = `${apiBaseURL}/Home/GetAllFacultyMembers/`
 
@@ -7,3 +8,5 @@ export const staffMembersURL = `${apiBaseURL}/Home/GetAllFacultyStaffMembers/`
 export const loginUserURL = `${apiBaseURL}/Login/LogonFromDevice/`
 
 export const registerUserURL = `${apiBaseURL}/Login/StudentRegister/`
+
+export const myCoursesURL = `${apiBaseURL}/Home/GetMyCourses/`
